phase2/requirements.txt
# ── Phase 1: Data Collection & Processing ─────────────────────────────────────
aiohttp==3.9.5
aiofiles==23.2.1
requests==2.31.0
tenacity==8.3.0
pandas==2.2.2
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1
orjson==3.10.3
pyarrow==16.1.0
typer==0.12.3
rich==13.7.1
tqdm==4.66.4

# ── Phase 2: Embeddings & Vector Stores ───────────────────────────────────────
sentence-transformers>=2.7,<4.0
chromadb>=0.5,<1.0
faiss-cpu>=1.8,<2.0
openai>=1.35,<2.0

# ── Phase 3 (upcoming): RAG Orchestration ─────────────────────────────────────
# langchain>=0.2,<0.3
# langchain-openai>=0.1,<0.2
# langchain-community>=0.2,<0.3

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7